pyproject.toml
[project]
name = "agentllm-agents-rhdh"
version = "1.0.0"
description = "RHDH-specific agents for AgentLLM"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "RHDH Team"}
]
keywords = ["agent", "llm", "rhdh", "release-manager", "jira"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "agentllm-core>=0.1.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.0.0",
    "jira>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
]

# Entry point registration for plugin discovery
[project.entry-points."agentllm.agents"]
release-manager = "agentllm_agents_rhdh.release_manager:ReleaseManagerFactory"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentllm_agents_rhdh"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
